requirements.txt
langchain
langchain-mistralai
mistralai
playwright
beautifulsoup4
python-dotenv
